Implementing authentication using JWT tokens and user access control

+++ src/main/java/com/example/university_system/controller/CourseController.java
@@ -1,10 +1,8 @@
 package com.example.university_system.controller;
 
-
-
 import com.example.university_system.controller.Base.BaseController;
+import com.example.university_system.enums.Permission;
 import com.example.university_system.service.impl.CourseService;
-
 import com.example.university_system.dto.course.AddCourseDTO;
 import com.example.university_system.dto.course.UpdateCourseDTO;
 import com.example.university_system.dto.course.ViewCourseDTO;
@@ -23,7 +21,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/course/v1/")
+@RequestMapping("${application_controllers_CourseController}")
 @RequiredArgsConstructor
 public class CourseController extends BaseController<
         CourseEntity, Long, AddCourseDTO, UpdateCourseDTO, ViewCourseDTO
@@ -44,15 +42,54 @@ public class CourseController extends BaseController<
         return courseMapper;
     }
 
+    @Override
+    protected Permission getCreatePermission() {
+        return Permission.COURSE_CREATE;
+    }
+
+    @Override
+    protected Permission getReadPermission() {
+        return Permission.COURSE_READ;
+    }
+
+    @Override
+    protected Permission getReadAllPermission() {
+        return Permission.COURSE_READ_ALL;
+    }
+
+    @Override
+    protected Permission getUpdatePermission() {
+        return Permission.COURSE_UPDATE;
+    }
+
+    @Override
+    protected Permission getDeletePermission() {
+        return Permission.COURSE_DELETE;
+    }
+
+    @Override
+    protected void checkMonoAndAllPermissionWithOwnership(Long resourceId, Permission permission, Permission allPermission) {
+        checkPermission(permission);
+    }
+
+    @Override
+    protected void checkMonoPermissionWithOwnership(Long resourceId, Permission permission) {
+        checkPermission(permission);
+    }
+
 
     @GetMapping("/code/{code}")
     public ViewCourseDTO findByCode(@PathVariable int code) {
+        this.checkPermission(Permission.COURSE_READ);
+
         return courseMapper.toViewDto(courseService.findByCode(code));
     }
 
     @GetMapping("/{codeCourse}/students")
     @ResponseStatus(HttpStatus.OK)
     public List<ViewStudentDTO> listStudents(@PathVariable int codeCourse) {
+        this.checkPermission(Permission.STUDENT_COURSE_READ);
+
         List<StudentEntity> studentEntities = courseService.listStudents(codeCourse);
         return studentMapper.toListViewDTO(studentEntities);
     }
@@ -60,29 +97,39 @@ public class CourseController extends BaseController<
     @PostMapping("/{codeCourse}/students/add/{stdNumber}")
     @ResponseStatus(HttpStatus.OK)
     public void addStudent(@PathVariable int codeCourse, @PathVariable long stdNumber) {
+        this.checkPermission(Permission.STUDENT_COURSE_CREATE);
+
         courseService.addStudent(codeCourse, stdNumber);
     }
 
     @DeleteMapping("/{codeCourse}/students/delete/{stdNumber}")
     @ResponseStatus(HttpStatus.OK)
     public void removeProfessor(@PathVariable int codeCourse, @PathVariable long stdNumber) {
+        this.checkPermission(Permission.STUDENT_COURSE_DELETE);
+
         courseService.removeStudent(codeCourse, stdNumber);
     }
 
     @PostMapping("/{codeCourse}/professor/assign/{codeProfessor}")
     @ResponseStatus(HttpStatus.OK)
     public void assignProfessor(@PathVariable int codeCourse, @PathVariable int codeProfessor) {
+        this.checkPermission(Permission.PROFESSOR_COURSE_CREATE);
+
         courseService.assignProfessor(codeCourse, codeProfessor);
     }
 
     @DeleteMapping("/{codeCourse}/professor/remove")
     @ResponseStatus(HttpStatus.OK)
     public void removeProfessor(@PathVariable int codeCourse) {
+        this.checkPermission(Permission.PROFESSOR_COURSE_DELETE);
+
         courseService.removeProfessor(codeCourse);
     }
 
     @GetMapping("/{codeCourse}/professor")
     public ViewProfessorDTO getProfessor(@PathVariable int codeCourse) {
+        this.checkPermission(Permission.PROFESSOR_COURSE_READ);
+
         ProfessorEntity professorEntity = courseService.getProfessor(codeCourse);
         return professorMapper.toViewDto(professorEntity);
     }
